Moved preferredStyleSchemes schema and properties to root (fixed Apple WWDC video misinformation)
